backends/none: Use dedicated SSH keypair.
This should make the "none" backend a bit more useful as we only need to
provide a key _or_ passphrase on machine creation and from thereon we
only use the SSH keys within NixOps' state database.

If you have a large set of "none" machines and want to export/import
them, this should no longer require to move around ssh_configs or even
entries in your hosts file.

In addition, this fixes the situation where you need to build on the
target machines and the none backend wasn't able to provide a private
key for the to be generated nix.machines file.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>