backends/none: Deploy SSH key via physical spec.
We now no longer create a .ssh/authorized_keys file, as suggested by
@edolstra in:

https://github.com/NixOS/nixops/pull/201#issuecomment-49752969

The key is created in the machine state and passed via the physical
specification and we're checking for the cur_toplevel in the machine
state to ensure that we've deployed already.

So until the first deployment, NixOps (or better OpenSSH) either asks
for a password or if there's a key available uses the external key.

That way, we can also deploy configurations with users.mutableUsers set
to false, because it requires to either have a password or a key set in
the configuration and otherwise won't build.

Of course, we also adapted the VM test to properly test for this change.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>